backend/requirements.txt
asgiref>=3.7.2
Django>=4.2
django-filter>=23.2
djangorestframework>=3.14
Pillow>=10.0.0
